Struct daosign_app::daosign_app::DAOsignApp

pub struct DAOsignApp { /* private fields */ }

Main storage structure for DAOsignApp contract.

Implementations

impl DAOsignApp

DAOsignApp contract implementation.

pub fn new(domain: EIP712Domain) -> Self

Ink! constructor for creating a new DAOsignApp instance.

This constructor initializes a new DAOsignApp contract instance with the provided EIP712 domain.

Arguments

domain - EIP712Domain struct representing the domain of the contract.

Returns

A new instance of DAOsignApp.

pub fn store_proof_of_authority(&mut self, data: SignedProofOfAuthority)

Ink! message to store a Proof of Authority.

This function stores a Proof of Authority along with its signature and validates the signature and message. If the data is valid, it is stored in the contract, and relevant mappings are updated.

Arguments

data - SignedProofOfAuthority struct containing the proof of authority data and its signature.

pub fn store_proof_of_signature(&mut self, data: SignedProofOfSignature)

Ink! message to store a Proof of Signature.

This function stores a Proof of Signature along with its signature and validates the signature and message. If the data is valid, it is stored in the contract, and relevant mappings are updated.

Arguments

data - SignedProofOfSignature struct containing the proof of signature data and its signature.

pub fn store_proof_of_agreement(&mut self, data: SignedProofOfAgreement)

Ink! message to store a Proof of Agreement.

This function stores a Proof of Agreement and validates the message. If the data is valid, it is stored in the contract.

Arguments

data - SignedProofOfAgreement struct containing the proof of agreement data.

pub fn get_proof_of_authority(&self, cid: String) -> SignedProofOfAuthority

Ink! message to retrieve a Proof of Authority by its CID.

This function retrieves a stored Proof of Authority by its CID.

Arguments

cid - String representing the CID of the Proof of Authority.

pub fn get_proof_of_signature(&self, cid: String) -> SignedProofOfSignature

Ink! message to retrieve a Proof of Signature by its CID.

This function retrieves a stored Proof of Signature by its CID.

Arguments

cid - String representing the CID of the Proof of Signature.

pub fn get_proof_of_agreement(&self, cid: String) -> SignedProofOfAgreement

Ink! message to retrieve a Proof of Agreement by its CID.

This function retrieves a stored Proof of Agreement by its CID.

Arguments

cid - String representing the CID of the Proof of Agreement.

pub fn validate_signed_proof_of_authority( &self, data: SignedProofOfAuthority ) -> bool

Validates a signed Proof-of-Authority message.

This function performs various checks on the provided SignedProofOfAuthority data to ensure its validity.

Arguments

data - SignedProofOfAuthority struct representing the signed proof.

Returns

Returns true if the validation passes, otherwise raises assertions.

pub fn validate_signed_proof_of_signature( &self, data: SignedProofOfSignature ) -> bool

Validates a signed Proof-of-Signature message.

This function performs various checks on the provided SignedProofOfSignature data to ensure its validity.

Arguments

data - SignedProofOfSignature struct representing the signed proof.

Returns

Returns true if the validation passes, otherwise raises assertions.

pub fn validate_signed_proof_of_agreement( &self, data: SignedProofOfAgreement ) -> bool

Validates a signed Proof-of-Agreement message.

This function performs various checks on the provided SignedProofOfAgreement data to ensure its validity.

Arguments

data - SignedProofOfAgreement struct representing the signed proof.

Returns

Returns true if the validation passes, otherwise raises assertions.